libtinyrad: adjusting datatypes in TinyRadMap to fix padding 32 bit archs

+++ lib/libtinyrad/libtinyrad.h
@@ -135,7 +135,7 @@ struct _tinyrad
 struct _tinyrad_map
 {
    const char *          name;
-   uint64_t              value;
+   uintptr_t             value;
 };
